Structure files and folder for ETL pipeline.
