+++ apps/etl/extraction/sources/base/extract.py
@@ -2,7 +2,7 @@ import requests
 from celery.utils.log import get_task_logger
 from django.core.exceptions import ObjectDoesNotExist
 
-from .models import ExtractionData
+from apps.etl.models import ExtractionData
 
 logger = get_task_logger(__name__)
