+++ apps/etl/management/commands/import_gdacs_data.py
@@ -2,8 +2,8 @@ import logging
 
 from django.core.management.base import BaseCommand
 
+from apps.etl.etl_tasks.gdacs import import_hazard_data
 from apps.etl.models import HazardType
-from apps.etl.tasks import import_hazard_data
 
 logger = logging.getLogger(__name__)
 
@@ -13,9 +13,9 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         import_hazard_data.delay("EQ", HazardType.EARTHQUAKE)
-        import_hazard_data.delay("TC", HazardType.CYCLONE)
-        import_hazard_data.delay("FL", HazardType.FLOOD)
-        import_hazard_data.delay("DR", HazardType.DROUGHT)
-        import_hazard_data.delay("WF", HazardType.WILDFIRE)
-        import_hazard_data.delay("VO", HazardType.VOLCANO)
-        import_hazard_data.delay("TS", HazardType.TSUNAMI)
+        # import_hazard_data.delay("TC", HazardType.CYCLONE)
+        # import_hazard_data.delay("FL", HazardType.FLOOD)
+        # import_hazard_data.delay("DR", HazardType.DROUGHT)
+        # import_hazard_data.delay("WF", HazardType.WILDFIRE)
+        # import_hazard_data.delay("VO", HazardType.VOLCANO)
+        # import_hazard_data.delay("TS", HazardType.TSUNAMI)
