+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "pydantic>=2.10.2,<3",
     "pystac-monty",
     "requests>=2.32.3",
+    "geopandas",
 ]
 
 [tool.uv.sources]
